Renames tslibPkg to tslib
