+++ src/apps/chifra/pkg/tslib/list.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a license that can
 // be found in the LICENSE file.
 
-package tslibPkg
+package tslib
 
 import (
 	"encoding/csv"
