+++ src/apps/chifra/pkg/tslib/tots.go
@@ -1,10 +1,4 @@
-// Package `tslibPkg` provides conversions between four entities: block numbers, dates, Linux
-// timestamps, and special named blocks. The function names are self-explanitory.
-//
-// Note: because the relationship between block numbers and dates or timestamps is a per-chain
-// value, some of these functions require a `chain` name. Any chain name is permissable, but
-// they all require proper configuration in the TrueBlocks config files.
-package tslibPkg
+package tslib
 
 import (
 	"strings"
