+++ src/apps/chifra/pkg/tslib/tslib_integration_test.go
@@ -5,7 +5,7 @@
 // Use of this source code is governed by a license that can
 // be found in the LICENSE file.
 
-package tslibPkg
+package tslib
 
 import (
 	"testing"
